backend/requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib 
httpx 
tortoise-orm[asyncpg] 
flask[async]
uvicorn
flask-cors